requirements.txt
Django>=5.1.0
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
python-decouple>=3.8
requests>=2.25.0
coreapi>=2.3.3
PyYAML>=6.0
uritemplate>=4.1.1
gunicorn>=21.2.0
whitenoise>=6.6.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
djangorestframework-simplejwt>=5.3.0
pillow>=10.0.0  # Para manejo de imágenes desde Flutter